+++ test/unit/object/test_search.py
@@ -12,14 +12,24 @@ def search_query():
     return 'myquery'
 
 
-@pytest.fixture
-def search_limit():
-    return 20
+@pytest.fixture(params=(1, 20, 100))
+def search_limit(request):
+    return request.param
 
 
-@pytest.fixture
-def search_offset():
-    return 0
+@pytest.fixture(params=(0, 10))
+def search_offset(request):
+    return request.param
+
+
+@pytest.fixture(params=(None, 'file', 'folder'))
+def search_result_type(request):
+    return request.param
+
+
+@pytest.fixture(params=(None, ('name',), ('name', 'description')))
+def search_content_types(request):
+    return request.param
 
 
 @pytest.fixture
@@ -81,6 +91,8 @@ def compare_params(self, other):
             if json.loads(self['mdfilters']) != json.loads(other['mdfilters']):
                 return False
         # For other keys, just ensure that they are equal
+        elif key in ('type', 'content_types'):
+            return self[key] is None or self[key] == other[key]
         else:
             if self[key] != other[key]:
                 return False
@@ -96,20 +108,31 @@ def test_search_with_value_based_filters(
         search_offset,
         search_value_based_filters,
         search_response,
-        search_entries
+        search_entries,
+        search_result_type,
+        search_content_types,
 ):
     # pylint:disable=redefined-outer-name
     mock_box_session.get.return_value, _ = make_mock_box_request(response=search_response)
-    response = test_search.search(search_query, limit=search_limit, offset=search_offset, metadata_filters=search_value_based_filters)
+    response = test_search.search(
+        search_query,
+        limit=search_limit,
+        offset=search_offset,
+        metadata_filters=search_value_based_filters,
+        result_type=search_result_type,
+        content_types=search_content_types,
+    )
     assert response == [File(mock_box_session, search_entry['id'], search_entry) for search_entry in search_entries]
 
     mock_box_session.get.assert_called_once_with(
         test_search.get_url(),
         params=Matcher(compare_params, {
-            'query': 'myquery',
-            'limit': 20,
+            'query': search_query,
+            'limit': search_limit,
             'mdfilters': json.dumps(search_value_based_filters.as_list()),
-            'offset': 0
+            'offset': search_offset,
+            'type': search_result_type,
+            'content_types': ','.join(search_content_types) if search_content_types else search_content_types,
         })
     )
 
@@ -123,20 +146,31 @@ def test_search_with_range_filters(
         search_offset,
         search_range_filters,
         search_response,
-        search_entries
+        search_entries,
+        search_result_type,
+        search_content_types,
 ):
     # pylint:disable=redefined-outer-name
     mock_box_session.get.return_value, _ = make_mock_box_request(response=search_response)
-    response = test_search.search(search_query, limit=search_limit, offset=search_offset, metadata_filters=search_range_filters)
+    response = test_search.search(
+        search_query,
+        limit=search_limit,
+        offset=search_offset,
+        metadata_filters=search_range_filters,
+        result_type=search_result_type,
+        content_types=search_content_types,
+    )
     assert response == [File(mock_box_session, search_entry['id'], search_entry) for search_entry in search_entries]
 
     mock_box_session.get.assert_called_once_with(
         test_search.get_url(),
         params=Matcher(compare_params, {
-            'query': 'myquery',
-            'limit': 20,
+            'query': search_query,
+            'limit': search_limit,
             'mdfilters': json.dumps(search_range_filters.as_list()),
-            'offset': 0
+            'offset': search_offset,
+            'type': search_result_type,
+            'content_types': ','.join(search_content_types) if search_content_types else search_content_types,
         })
     )